+++ flogin/default_events.py
@@ -28,6 +28,8 @@ def get_default_events(plugin: Plugin) -> dict[str, Callable[..., Awaitable[Any]
     def on_query(data: dict[str, Any], raw_settings: dict[str, Any]):
         query = Query(data)
         if plugin._settings_are_populated is False:
+            LOG.info(f"Settings have not been populated yet, creating a new instance")
+            plugin._settings_are_populated = True
             plugin.settings = Settings(raw_settings)
         else:
             plugin.settings._update(raw_settings)